Some cool refactoring and improvements to the script.

+++ create_ssh_ramdisk.py
@@ -20,6 +20,8 @@ def run_cmd(cmd):
     subp = subprocess.Popen(cmd, shell=True, stdout=subprocess.PIPE)
     return subp.stdout.read().decode('utf-8')
 
+# NOTE: This is not my code.
+# Had to pluck them out from 'autodecrypt' module and do some very slight modifications on a few of the functions.
 # region decrypt_img4 functions
 
 
@@ -43,7 +45,8 @@ def get_image_type(filename: str):
 
 
 def decrypt_img(infile: str, magic: str, iv: str, key: str) -> int:
-    """Decrypt IM4P file. This code is mostly copy/pasta from PyIMG4."""
+    """Decrypt IM4P file. This code is mostly copy/pasta from decrypt_img4.py
+       which itself is a copy/pasta from PyIMG4."""
     file = open(infile, 'rb').read()
     im4p = pyimg4.IM4P(file)
 
@@ -68,10 +71,14 @@ def decrypt_img(infile: str, magic: str, iv: str, key: str) -> int:
 
     open(outfile, 'wb').write(im4p.payload.output().data)
     return 0
+
+
 # endregion
 
 def run_pcmd(cmd):
-    print(run_cmd(cmd))
+    ret = run_cmd(cmd).strip()
+    if ret:
+        print(ret)
 
 
 def clean_up():
@@ -95,9 +102,8 @@ def get_gaster(platform):
         print(f"[!] Failed to download gaster for platform {platform}!")
 
 
-def get_ipsw_url(product_type, ios_version):
-    global ipsw_url
-    ipsw_url = 'null'
+def get_url_and_build_id(product_type, ios_version):
+    ipsw_url = None
     api_path = f'https://api.ipsw.me/v4/device/{product_type}?type=ipsw'
     resp = requests.get(api_path)
     if resp.status_code == 200 and resp.text:
@@ -105,12 +111,14 @@ def get_ipsw_url(product_type, ios_version):
         for firmware_list in resp['firmwares']:
             if firmware_list['version'] == ios_version:
                 ipsw_url = firmware_list['url']
-    return ipsw_url
+                build_id = firmware_list['buildid']
+    return ipsw_url, build_id if ipsw_url is not None and build_id is not None else None
 
 
 # kerneldiff.py
 # https://github.com/verygenericname/SSHRD_Script/blob/main/kerneldiff.py
 def kernel_diff(original, patched, bpatchfile):
+    print("[*] Comparing patched kernel with original...")
     sizeP = os.path.getsize(patched)
     sizeO = os.path.getsize(original)
     if sizeP != sizeO:
@@ -131,33 +139,24 @@ def kernel_diff(original, patched, bpatchfile):
         print(data)
 
 
-def get_build_id(product_type, ios_version):
-    global build_id
-    build_id = 'null'
-    api_path = f'https://api.ipsw.me/v4/device/{product_type}?type=ipsw'
-    resp = requests.get(api_path)
-    if resp.status_code == 200 and resp.text:
-        resp = resp.json()
-        for firmware_list in resp['firmwares']:
-            if firmware_list['version'] == ios_version:
-                build_id = firmware_list['buildid']
-    return build_id
-
-
 def download_required_files():
-    ipsw_url = get_ipsw_url(args.product_type, args.ios)
+    ipsw_url, build_id = get_url_and_build_id(args.product_type, args.ios)
+    if ipsw_url is None or build_id is None:
+        print(f'[!] Required files are not found for this version of iOS!')
+        exit(1)
     print(f'URL: {ipsw_url}')
     run_pcmd(f"../{sys_platform}/pzb -g BuildManifest.plist {ipsw_url}")
     with open('BuildManifest.plist', 'rb') as build_manifest:
         build_manifest_plist = plistlib.load(build_manifest)
-        codename: str = build_manifest_plist['BuildIdentities'][0]['Info']['BuildTrain']
-        ibss_path: str = build_manifest_plist['BuildIdentities'][0]['Manifest']['iBSS']['Info']['Path']
-        ibec_path: str = build_manifest_plist['BuildIdentities'][0]['Manifest']['iBEC']['Info']['Path']
-        devicetree_path: str = build_manifest_plist['BuildIdentities'][0]['Manifest']['DeviceTree']['Info']['Path']
+        build_identities = build_manifest_plist['BuildIdentities']
+        codename: str = build_identities[0]['Info']['BuildTrain']
+        ibss_path: str = build_identities[0]['Manifest']['iBSS']['Info']['Path']
+        ibec_path: str = build_identities[0]['Manifest']['iBEC']['Info']['Path']
+        devicetree_path: str = build_identities[0]['Manifest']['DeviceTree']['Info']['Path']
 
-        restoreramdisk_path: str = build_manifest_plist['BuildIdentities'][0]['Manifest']['RestoreRamDisk']['Info'][
+        restoreramdisk_path: str = build_identities[0]['Manifest']['RestoreRamDisk']['Info'][
             'Path']
-        kernelcache_path: str = build_manifest_plist['BuildIdentities'][0]['Manifest']['RestoreKernelCache']['Info'][
+        kernelcache_path: str = build_identities[0]['Manifest']['RestoreKernelCache']['Info'][
             'Path']
         trustcache_path: str = f'Firmware/{restoreramdisk_path}.trustcache'
 
@@ -174,42 +173,46 @@ def download_required_files():
         devicetree_path = devicetree_path.replace("Firmware/dfu/", "").replace("Firmware/all_flash/", "")
         trustcache_path = trustcache_path.replace("Firmware/", "")
 
-        return ibss_path, ibec_path, kernelcache_path, restoreramdisk_path, trustcache_path, devicetree_path
+        return ibss_path, ibec_path, kernelcache_path, restoreramdisk_path, trustcache_path, devicetree_path, build_id
 
 
-def decrypt_boot_stages(ibss_path, ibec_path):
-    print(f"Decryption mode: { 'Gaster' if args.decrypt_mode == 1 else 'Online key fetch'}")
+# Wrapper around decrypt_img
+def decrypt_img4p(infile, buildid):
+    print(f'[*] Decrypting: "{infile}"...')
+    ivkey = scrapkeys.getkeys(args.product_type, buildid, infile)
+    if ivkey is None:
+        print(f'[!] Internal error! Failed to get IvKey for BuildID: "{buildid}". Aborting...')
+        exit(1)
+    iv, key = utils.split_key(ivkey)
+    magic = get_image_type(infile)
+    print(f'[*] IV: "{iv}"')
+    print(f'[*] Key: "{key}"')
+    print(f'[*] Magic: "{magic}"')
+    return True if decrypt_img(infile, magic, iv, key) == 0 else False
+
+
+
+def decrypt_boot_stages(ibss_path, ibec_path, build_id):
+    print(f"Decryption mode: {'Gaster' if args.decrypt_mode == 1 else 'Online key fetch'}")
+    ibec_decrypted_path = ibec_path.replace(".im4p", ".bin")
+    ibss_decrypted_path = ibss_path.replace(".im4p", ".bin")
+
     if args.decrypt_mode == 1:
-        ibec_decrypted_path = ibec_path.replace(".im4p", ".bin")
-        ibss_decrypted_path = ibss_path.replace(".im4p", ".bin")
         run_pcmd(f'../{sys_platform}/gaster pwn')
         run_pcmd(f'../{sys_platform}/gaster reset')
         run_pcmd(f'../{sys_platform}/gaster decrypt {ibss_path} {ibss_decrypted_path}')
         run_pcmd(f'../{sys_platform}/gaster decrypt {ibec_path} {ibec_decrypted_path}')
     else:
-        build_id = get_build_id(args.product_type, args.ios)
         print(f'[*] Build ID: {build_id}')
-        print(f'[*] Product Type: {args.product_type}\n')
+        print(f'[*] Product Type: {args.product_type}')
+        print("[*] Reached iBSS & iBEC decryption stage!")
+        if not decrypt_img4p(ibss_path, build_id):
+            print('[!] Failed to decrypt iBSS! Aborting...')
+            exit(1)
+        if not decrypt_img4p(ibec_path, build_id):
+            print('[!] Failed to decrypt iBEC! Aborting...')
+            exit(1)
 
-        print("[!] Beginning decryption of iBSS and iBEC...\n")
-        print(f'[*] iBSS Path: {ibss_path}')
-        ivkey_ibss = scrapkeys.getkeys(args.product_type, build_id, ibss_path)
-        iv_ibss, key_ibss = utils.split_key(ivkey_ibss)
-        print(f'[*] IV: {iv_ibss}')
-        print(f'[*] Key: {key_ibss}')
-        magic_ibss = get_image_type(ibss_path)
-        print(ibss_path, magic_ibss, iv_ibss, key_ibss)
-        decrypt_img(ibss_path, magic_ibss, iv_ibss, key_ibss)
-
-        print(f'[*] iBEC Path: {ibec_path}')
-        ivkey_ibec = scrapkeys.getkeys(args.product_type, build_id, ibec_path)
-        iv_ibec, key_ibec = utils.split_key(ivkey_ibec)
-        print(f'[*] IV: {iv_ibec}')
-        print(f'[*] Key: {key_ibec}')
-        magic_ibec = get_image_type(ibec_path)
-        decrypt_img(ibec_path, magic_ibec, iv_ibec, key_ibec)
-        ibec_decrypted_path = ibec_path.replace(".im4p", ".bin")
-        ibss_decrypted_path = ibss_path.replace(".im4p", ".bin")
     return ibss_decrypted_path, ibec_decrypted_path
 
 
@@ -225,7 +228,7 @@ def patch_files(ibss_decrypted_path, ibec_decrypted_path, kernelcache_path, devi
     if args.cpid == "0x8960" or args.cpid == "0x7000" or args.cpid == "0x7001":
         boot_args = boot_args[:-1]
         boot_args += ' -restore"'
-    print(f"[!] Boot arguments: {boot_args}")
+    print(f"[*] Boot arguments: {boot_args}")
     run_pcmd(f'../{sys_platform}/iBoot64Patcher {ibec_decrypted_path} iBEC.patched -b {boot_args} -n')
     run_pcmd(
         f'../{sys_platform}/img4 -i iBEC.patched -o ../final_ramdisk/{args.ios}/{args.product_type}/{args.model}/iBEC.img4 -M IM4M -A -T ibec')
@@ -241,12 +244,13 @@ def patch_files(ibss_decrypted_path, ibec_decrypted_path, kernelcache_path, devi
     run_pcmd(
         f'../{sys_platform}/img4 -i {trustcache_path} -o ../final_ramdisk/{args.ios}/{args.product_type}/{args.model}/trustcache.img4 -M IM4M -T rtsc')
     run_pcmd(f'../{sys_platform}/img4 -i {ramdisk_path} -o ramdisk.dmg')
-    patch_ramdisk('ramdisk.dmg')
+    patch_ramdisk()
     run_pcmd(
         f'../{sys_platform}/img4 -i ../other/logo.im4p -o ../final_ramdisk/{args.ios}/{args.product_type}/{args.model}/logo.img4 -A -T rlgo -M IM4M')
 
 
-def patch_ramdisk(ramdisk_path):
+def patch_ramdisk():
+    print("[*] Patching ramdisk...")
     if sys_platform == 'Darwin':
         run_pcmd('hdiutil resize -size 210MB ramdisk.dmg')
         run_pcmd('hdiutil attach -mountpoint /tmp/SSHRD ramdisk.dmg')
@@ -268,7 +272,7 @@ def patch_ramdisk(ramdisk_path):
             run_pcmd(f'../{sys_platform}/hfsplus ramdisk.dmg untar ../sshtars/atvssh.tar > /dev/null')
         elif args.cpid == '0x8012':
             run_pcmd(f'../{sys_platform}/hfsplus ramdisk.dmg untar ../sshtars/t2ssh.tar > /dev/null')
-            print("[!] !!! T2 Ssh might hang and do nothing when booting !!!")
+            print("[!] !!! T2 SSH might hang and do nothing when booting !!!")
         else:
             run_pcmd(f'../{sys_platform}/hfsplus ramdisk.dmg untar ../sshtars/ssh.tar > /dev/null')
     run_pcmd(
@@ -295,6 +299,13 @@ if __name__ == '__main__':
     parser.add_argument('--boot_args', '-ba', type=str,
                         help='iOS arguments to execute during boot. Default: "rd=md0 debug=0x2014e -v wdt=-1"')
     args = parser.parse_args()
+    # People might mistakenly use commas in the iOS version.
+    if args.ios is not None and ',' in args.ios:
+        args.ios = args.ios.replace(',', '.')
+
+    if not path.isfile(f'other/shsh/{args.cpid}.shsh'):
+        print(f'[!] CPID ({args.cpid}) is not found or is not supported.')
+        exit(1)
 
     if not os.path.isfile(f'{sys_platform}/gaster') and args.decrypt_mode == 1:
         print("[!] gaster does not appear to exist! Downloading a new one...\n")
@@ -305,21 +316,28 @@ if __name__ == '__main__':
     pathlib.Path(f'final_ramdisk/{args.ios}/{args.product_type}').mkdir(exist_ok=True, parents=True)
 
     if path.isdir(f'final_ramdisk/{args.ios}/{args.product_type}/{args.model}'):
-        print(f'[!] data for {args.model} already exists! Recreating...')
+        choice = input(f'[!] Data for {args.product_type} ({args.model}/{args.ios}) already exists! Do you want to delete them and start over? (y/N): ') or 'N'
+        if choice.lower() == 'n':
+            print('[*] Exiting...')
+            exit(0)
         shutil.rmtree(f'final_ramdisk/{args.ios}/{args.product_type}/{args.model}')
+    files = glob.glob('temp_ramdisk/*')
+    for file in files:
+        os.remove(file)
 
     pathlib.Path(f'final_ramdisk/{args.ios}/{args.product_type}/{args.model}').mkdir(exist_ok=True, parents=True)
 
     os.chdir('temp_ramdisk')
     run_pcmd(f"../{sys_platform}/img4tool -e -s ../other/shsh/{args.cpid}.shsh -m IM4M")
 
-    ibss_path, ibec_path, kernelcache_path, restoreramdisk_path, trustcache_path, devicetree_path = download_required_files()
-    ibss_decrypted_path, ibec_decrypted_path = decrypt_boot_stages(ibss_path, ibec_path)
+    ibss_path, ibec_path, kernelcache_path, restoreramdisk_path, trustcache_path, devicetree_path, build_id = download_required_files()
+    ibss_decrypted_path, ibec_decrypted_path = decrypt_boot_stages(ibss_path, ibec_path, build_id)
     patch_files(ibss_decrypted_path, ibec_decrypted_path, kernelcache_path, devicetree_path, restoreramdisk_path,
                 trustcache_path)
 
     os.chdir('../')
     clean_up()
+    print(f"[*] Ramdisk files saved to: {main_root_dir}/final_ramdisk/{args.ios}/{args.product_type}/{args.model}")
     print("[*] Done!")
     print("Python version was made by Bonkeyzz.")
-    print("Original Shell script was made by verygenericname.")
+    print("Original Shell script was made by verygenericname (https://github.com/verygenericname).")
